requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pydantic[email]>=2.8.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0,<6.0.0
python-multipart>=0.0.9
alembic>=1.13.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
